pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-news-feishu"
version = "0.1.0"
description = "AI新闻推送飞书工具 - 自动获取AI新闻并推送到飞书"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
keywords = ["ai", "news", "feishu", "rss", "automation"]
dependencies = [
    "feedparser==6.0.10",
    "requests==2.31.0",
    "lark-oapi==1.2.0",
    "PyYAML==6.0",
    "python-dotenv==1.0.0",
    "python-dateutil==2.8.2",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.0",
    "black==23.11.0",
    "flake8==6.1.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/ai-news-feishu"
Repository = "https://github.com/yourusername/ai-news-feishu"
Issues = "https://github.com/yourusername/ai-news-feishu/issues"

[project.scripts]
ai-news-feishu = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]